Fix updating kits failing when going through upsert flow
The issue here was that we where connecting kit options in the upsert,
without removing the existing options. This was causing duplicate options
on the specific slot keys.

We'll be looking into updating the db schema to index on

```
kitbase id, option id, and slotkey
```

so that the db will yell at us if something like this happens in the future.